Add screenrc

+++ Makefile
@@ -1,2 +1,3 @@
 links:
-	ln -s ${PWD}/tmux.conf ~/.tmux.conf
+	[ -f ~/.tmux.conf ] || ln -s ${PWD}/tmux.conf ~/.tmux.conf
+	[ -f ~/.screenrc ] || ln -s ${PWD}/screenrc ~/.screenrc
